# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "0.1"
    },
    "paths": {
        "api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Retrive a listing of the Companies",
                "operationId": "a8122b7dd94304bebfeded6578f13a93",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Store new Company",
                "operationId": "d6fd101fe893457e049dd69860c0db42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/companies/{uuid}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get an existing Company",
                "operationId": "a0cbc740f412d6e40a34f3a20bd0324d",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update an existing Company",
                "operationId": "2085ef4aa1358df5e5927f4230be0dcc",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete existing Company",
                "operationId": "ce4649ceb3fe11bbe9d80c237c34d8d9",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompanyRequest": {
                "title": "CompanyRequest",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "email": {
                        "type": "string",
                        "example": "example"
                    },
                    "address": {
                        "type": "string",
                        "example": "example"
                    },
                    "vat_number": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_name": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_email": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
            "Company": {
                "title": "Company",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                    },
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "email": {
                        "type": "string",
                        "example": "example"
                    },
                    "address": {
                        "type": "string",
                        "example": "example"
                    },
                    "vat_number": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_name": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "example"
                    },
                    "contact_email": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
        }
    }
}